requirements.txt
streamlit==1.51.0
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==12.0.0
face_recognition==1.3.0
face-recognition-models==0.3.0
dlib-bin==20.0.0

